Brain-Tumor/backend/requirements.txt
Flask==2.0.1
flask_cors
pillow
scikit-learn
opencv-python==4.5.3.56
numpy==1.21.2
scikit-image==0.18.3
Werkzeug==2.2.2